Revert "Fully support to multiple coins. Tested!"
