--- voltixApp/VoltixApp/Chains/EthereumHelper.swift
@@ -1,71 +1,0 @@
-import Foundation
-import Tss
-import WalletCore
-
-    // Define a custom error to indicate unimplemented methods
-enum EthereumHelperError: Error {
-    case notImplemented
-    case methodNotApplicable(String)
-}
-
-class EthereumHelper: CoinHelperProtocol {
-    func getSigningInput(utxos: [UtxoInfo], fromAddress: String, toAddress: String, toAmount: Int64, byteFee: Int64, memo: String?) -> Result<WalletCore.BitcoinSigningInput, Error> {
-            // Example implementation or throw not implemented error if not relevant
-        return .failure(EthereumHelperError.notImplemented)
-    }
-    
-    func getTransactionPlan(utxos: [UtxoInfo], fromAddress: String, toAddress: String, toAmount: Int64, byteFee: Int64, memo: String?) -> Result<WalletCore.BitcoinTransactionPlan, Error> {
-            // Example implementation or throw not implemented error if not relevant
-        return .failure(EthereumHelperError.notImplemented)
-    }
-    
-    func validateAddress(_ address: String) -> Bool {
-            // Implement Ethereum address validation logic
-        return CoinType.ethereum.validate(address: address)
-    }
-    
-    func getSignatureFromTssResponse(tssResponse: TssKeysignResponse) -> Result<Data, Error> {
-            // Example implementation or throw not implemented error if not relevant
-        return .failure(EthereumHelperError.notImplemented)
-    }
-    
-    func getCoinDetails(hexPubKey: String, hexChainCode: String) -> Result<Coin, Error> {
-            // Implement logic to derive Ethereum address and return Coin details or throw not implemented error
-        return .failure(EthereumHelperError.notImplemented)
-    }
-    
-    func getPublicKey(hexPubKey: String, hexChainCode: String) -> String {
-            // Implement logic to derive the public key or return an example/default value
-        return ""
-    }
-    
-    func getAddressFromPublicKey(hexPubKey: String, hexChainCode: String) -> Result<String, Error> {
-            // Implement logic to derive Ethereum address from public key or throw not implemented error
-        return .failure(EthereumHelperError.notImplemented)
-    }
-    
-    func getPreSigningImageHash(utxos: [UtxoInfo], fromAddress: String, toAddress: String, toAmount: Int64, byteFee: Int64, memo: String?) -> Result<[String], Error> {
-            // Since Ethereum doesn't use UTXOs, return method not applicable error
-        return .failure(EthereumHelperError.methodNotApplicable("Ethereum does not use UTXOs."))
-    }
-    
-    func getSigningInput(utxos: [UtxoInfo], fromAddress: String, toAddress: String, toAmount: Int64, byteFee: Int64, memo: String?) -> Result<Any, Error> {
-            // Throw not implemented error or adapt to Ethereum-specific input for transactions
-        return .failure(EthereumHelperError.notImplemented)
-    }
-    
-    func getPreSigningInputData(utxos: [UtxoInfo], fromAddress: String, toAddress: String, toAmount: Int64, byteFee: Int64, memo: String?) -> Result<Data, Error> {
-            // Since Ethereum transaction preparation differs, implement accordingly or throw not implemented error
-        return .failure(EthereumHelperError.notImplemented)
-    }
-    
-    func getTransactionPlan(utxos: [UtxoInfo], fromAddress: String, toAddress: String, toAmount: Int64, byteFee: Int64, memo: String?) -> Result<Any, Error> {
-            // Ethereum transactions don't use a plan in the same way as Bitcoin, adjust or throw error
-        return .failure(EthereumHelperError.notImplemented)
-    }
-    
-    func getSignedTransaction(utxos: [UtxoInfo], hexPubKey: String, fromAddress: String, toAddress: String, toAmount: Int64, byteFee: Int64, memo: String?, signatureProvider: (Data) -> Data) -> Result<String, Error> {
-            // Implement Ethereum transaction signing logic or throw not implemented error
-        return .failure(EthereumHelperError.notImplemented)
-    }
-}
